# docs.api.swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MadeAEat-Api",
    "description": "Api for the whole project",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://api-madeaeat.vercel.app/api/v1",
      "variables": {}
    },
    {
      "url": "http://localhost:8000/api/v1",
      "variables": {}
    }
  ],
  "paths": {
...
    "/client/commandes": {
      "get": {
        "tags": [
          "Commande - Users"
        ],
        "summary": "Get all commandes",
        "operationId": "Getallcommandes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}